requirements.txt
easygui~=0.98.3
pillow~=11.1.0